requirements.txt
tqdm==4.67.1
numpy==2.2.2
torch==2.5.1
open-clip-torch==2.23.0
huggingface-hub==0.27.1
transformers==4.48.0
scikit-learn==1.6.1
pandas==2.2.3
pillow==10.3.0
matplotlib==3.9.0
tabulate
